requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.5.0
pydantic-settings==2.1.0
pypdf==4.0.1
google-generativeai==0.3.0
python-dotenv==1.0.0
jinja2>=3.0
aiofiles>=23.1.0
requests>=2.28.0
google-auth>=2.17.3